Mostly finish scripting api, fully migrate main temp teleop to script

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/competition/utils/interactions/items/StandardDistanceSensor.java
@@ -22,7 +22,7 @@ public class StandardDistanceSensor extends InteractionSurface {
 
     public int getDistance(DistanceUnit unit) {
         SENSOR.setUnits(unit);
-        return SENSOR.getData();
+        return (int) SENSOR.getData();
     }
 
     public void stop() {
